Move FrequencyCappedAction to common package
